move all inverted indexes to one module
Also, remove `old_mmap_postings` and `compressed_postings`
